feature: add new script to macos os for develop on alis

+++ .scripts/macos/restart_nuget.sh
@@ -1,29 +1,42 @@
 #!/bin/bash
 
-cd ../../
-
-rm -rf ./.nuget/
-echo "./.nuget/"
-
-rm -rf ./**/obj/
-echo "./**/obj/"
-
-rm -rf ./**/bin/
-echo "./**/bin/"
-
-skip="Template"
-for i in `find . -name "*.csproj" -type f`; do
-    if [[ $i == *$skip* ]] ; then
-        echo "Skip project $i"
-    else
-        echo "Write default value of csproj = $i"
-        cat ./.config/Default_csproj.props > $i
-    fi
+echo "Do you want to continue? (y/n)"
+select yn in "Yes" "No"; do
+    case $yn in
+        Yes ) 
+          
+          cd ../../
+          
+          rm -rf ./.nuget/
+          echo "./.nuget/"
+          
+          rm -rf ./**/obj/
+          echo "./**/obj/"
+          
+          rm -rf ./**/bin/
+          echo "./**/bin/"
+          
+          skip="Template"
+          for i in `find . -name "*.csproj" -type f`; do
+              if [[ $i == *$skip* ]] ; then
+                  echo "Skip project $i"
+              else
+                  echo "Write default value of csproj = $i"
+                  cat ./.config/Default_csproj.props > $i
+              fi
+          done
+          
+          for i in `find . -name "*.csproj" -type f`; do
+              echo "restoring csproj = $i"
+              dotnet restore $i
+          done
+          
+          cd ./.scripts/macos/ || exit
+          
+          break;;
+        No ) 
+          echo "Goodbye!"
+          exit;;
+    esac
 done
 
-for i in `find . -name "*.csproj" -type f`; do
-    echo "restoring csproj = $i"
-    dotnet restore $i
-done
-
-cd ./.scripts/macos/ || exit
